+++ CHANGES.rst
@@ -13,6 +13,7 @@ Development
 - Fix purging of sensible configuration keys
 - Improve MAC address formatting when logging network status
 - Try 11 dB attenuation for measuring vcc
+- Disable heartbeat through RGB-LED, just blink twice on startup
 
 
 2019-06-17 0.4.0